requirements.txt
openai>=1.0.0
sentence-transformers>=2.7.0
xgboost>=2.0.0
optuna>=3.4.0
optuna-integration[xgboost]
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
python-dotenv>=1.0.0
joblib>=1.3.0
tqdm>=4.65.0
